requirements.txt
flask==3.0.0
flask-cors==4.0.0
pyyaml==6.0.1
colorama==0.4.6
rich==13.7.0
requests==2.31.0

